wishlist feature

+++ src/main/java/com/vichika/ecommercesystem/product/service/impl/WishlistServiceImpl.java
@@ -1,5 +1,11 @@
 package com.vichika.ecommercesystem.product.service.impl;
 
+import com.vichika.ecommercesystem.auth.model.AppUser;
+import com.vichika.ecommercesystem.cart.dto.response.CartResponse;
+import com.vichika.ecommercesystem.cart.model.Cart;
+import com.vichika.ecommercesystem.cart.model.CartItem;
+import com.vichika.ecommercesystem.cart.repository.CartItemRepository;
+import com.vichika.ecommercesystem.cart.repository.CartRepository;
 import com.vichika.ecommercesystem.common.PageResponse;
 import com.vichika.ecommercesystem.common.SortResponse;
 import com.vichika.ecommercesystem.exceptions.BadRequestException;
@@ -14,23 +20,29 @@ import com.vichika.ecommercesystem.product.repository.WishlistRepository;
 import com.vichika.ecommercesystem.product.service.WishlistService;
 import com.vichika.ecommercesystem.spec.SpecificationBuilder;
 import com.vichika.ecommercesystem.util.AuthUtil;
+import com.vichika.ecommercesystem.util.CartUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class WishlistServiceImpl implements WishlistService {
 
     private final WishlistRepository wishlistRepository;
     private final ProductRepository productRepository;
     private final WishlistMapper wishlistMapper;
+    private final CartRepository cartRepository;
+    private final CartItemRepository cartItemRepository;
     private final AuthUtil authUtil;
+    private final CartUtil cartUtil;
 
     @Override
     public WishlistResponse createWishlist(WishlistRequest request) {
@@ -49,6 +61,7 @@ public class WishlistServiceImpl implements WishlistService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public PageResponse<WishlistResponse> getMyWishlist(Long productId, String sortBy, String sortAs, Integer page, Integer size) {
 
         var user = authUtil.getCurrentUser();
@@ -70,15 +83,68 @@ public class WishlistServiceImpl implements WishlistService {
 
         var user = authUtil.getCurrentUser();
         var product = getProductById(productId);
-        var wishlist = wishlistRepository.findByUserAndProduct(user, product)
-                .orElseThrow(() -> new ResourceNotFoundException("Wishlist item not found"));
+        var wishlist = getUserAndProduct(user,product);
+        wishlistRepository.delete(wishlist);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public long count() {
+        var user = authUtil.getCurrentUser();
+        return wishlistRepository.countByUser(user);
+    }
+
+    @Override
+    public CartResponse moveToCart(WishlistRequest request) {
+
+        var user = authUtil.getCurrentUser();
+        var product = getProductById(request.productId());
+
+        var wishlist = getUserAndProduct(user,product);
+
+        if (product.getQuantity() <= 0) {
+            throw new BadRequestException("Product out of stock");
+        }
+
+        var cart = cartRepository.findByUser(user)
+                .orElseGet(() -> {
+                    var newCart = Cart.builder()
+                            .user(user)
+                            .build();
+                    return cartRepository.save(newCart);
+                });
+
+        var cartItem = cartItemRepository.findByCartAndProduct(cart, product)
+                .orElseGet(() -> {
+                    var item = CartItem.builder()
+                            .cart(cart)
+                            .product(product)
+                            .quantity(0)
+                            .build();
+                    return cartItemRepository.save(item);
+                });
+
+        if (cartItem.getQuantity() + 1 > product.getQuantity()) {
+            throw new BadRequestException("Insufficient stock");
+        }
+
+        cartItem.setQuantity(cartItem.getQuantity() + 1);
+
+        cartItemRepository.save(cartItem);
 
         wishlistRepository.delete(wishlist);
+
+        return cartUtil.buildCartResponse(cart);
     }
 
     private Product getProductById(Long productId){
         return productRepository.findById(productId)
                 .orElseThrow(() -> new ResourceNotFoundException("Product not found with id " + productId));
+    }
+
+    private Wishlist getUserAndProduct(AppUser user, Product product){
+        return wishlistRepository.findByUserAndProduct(user, product)
+                .orElseThrow(() -> new ResourceNotFoundException("Wishlist item not found"));
     }
 
 }
